Added Winner Overlay
When the number of correct matches is equal to half the number of cards
(all cards matched), an overlay appears to display the number of turns
and length of time it took to win the game.

Simple overlay with a title and text stats added in HTML. JS triggers
CSS transition to display the overlay and populates the text stats.

Signed-off-by: Avery Uslaner <uslaner.avery@gmail.com>